Extracted out getItems to reside in model/Subject
